AtomicContainers: Use 64-bit CAS timing counters
`CASStats` stored nanosecond counters as `UInt`. On 32-bit systems,
`time_ns()` produces `UInt64` deltas, so atomic updates failed with a
`UInt32`/`UInt64` type mismatch.

Store `total_ns` and `f_ns` as `UInt64`, matching `time_ns()` and the
other statistics types while avoiding rapid 32-bit counter overflow.

+++ src/AtomicContainers/AtomicContainers.jl
@@ -317,9 +317,9 @@ mutable struct CASStats <: AtomicStats
     "Maximum number of retries observed in any single `store!` call"
     @atomic max_retries::Int
     "Total time spent in all `store!` operations (nanoseconds)"
-    @atomic total_ns::UInt
+    @atomic total_ns::UInt64
     "Total time spent executing the user function `f(old)` (nanoseconds)"
-    @atomic f_ns::UInt
+    @atomic f_ns::UInt64
     CASStats() = new(0, 0, 0, 0, 0)
 end
